requirements.txt
aiogram>=3.7,<4.0
APScheduler>=3.10,<4.0
SQLAlchemy>=2.0,<3.0
aiosqlite>=0.20,<1.0
pydantic>=2.7,<3.0
pydantic-settings>=2.2,<3.0
python-dotenv>=1.0,<2.0
pytest>=8.0,<9.0
pytest-asyncio>=0.23,<1.0
